+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## develop branch
 
-Nothing yet.
+### New
+
+* Requirements\RequireAssignable - throw exception if an item isn't an assignable type
 
 ## 2.4.0 - Sat Jul 18 2015